Download this code
Open the code on your PC and compile it before uploading it on to your MCU.
Please check that the 'Board' and 'PORT' have been appropriatly selected. For this project the 'Board' will be Arduino Uno and the available port will be suggested by the IDE.
Select the 'Board' before compiling and as well as before uploading as the IDE tends to refresh this selection after compling. You have to select the suggested port just once, when you are uploading your code to the MCU